+++ implementation/core/src/main/java/io/smallrye/faulttolerance/core/circuit/breaker/NaiveRollingWindow.java
@@ -5,7 +5,7 @@ final class NaiveRollingWindow implements RollingWindow {
     private final int failureThreshold;
 
     private int index = 0;
-    private int counter = 0;
+    private long counter = 0; // long to avoid int overflow
 
     @SuppressWarnings("UnnecessaryThis")
     NaiveRollingWindow(int size, int failureThreshold) {
